Create: Newfile for AppConstants; Fix: Calling of Image and Color to Case
